Fixed Elevators - THANK THA LORD
It's been a bit chaotic but I ended up improving the workflow for new elevators.
Now it's not using an AnimationPlayer anymore, instead it uses Tweens to animate the platform at runtime. This way, it's possible to modify every instance of Elevator's animations just by modifying the StartingPosition and EndingPosition nodes.
Cool beans.